Create week 1 Question 1
Given an array of nonnegative integers, design a linear algorithm and implement it using a program
to find whether given key element is present in the array or not. Also, find total number of
comparisons for each input case. (Time Complexity = O(n), where n is the size of input)